Restructuring of code for cropping based on a uniform background
In order to make it easier to extend the functionality later, divided
the logic into two parts:
1. The "cropper" which does the actual cropping of the image based on
where the product is located in the image, the aspect ratio and
padding.
2. The "product finder" which handles the complicated part; figuring
out where the product is located in the image. Per now have only
implemented one function of this kind, the one for a uniform, single
colored background.

Also moved code around and into smaller functions.