Fix issue #2 : Handle a product at the edges of a picture
Suggestion for a (semi-) fix of issue #2 :

This solution is strict about the aspect ratio, so the images always
come out in 1500:1814. The way it does this is by adding padding to
the images, to enforce the aspect ratio. This padding is per now
white, which means the current solution only works for images with a
uniform white background. To make this padding look as natural as
possible some logic is needed to decide which edges the product is
touching, and then add padding only to the OTHER edges. There is
likely some untested edge cases where this will break.

This new solution seems to work nicely with all the
`Adidas-Advantage-Sneakers_484619_10` images as well as with the
`Adidas-Terrex-Trail-LS-T-shirt_484421_48` images. Have not tested any
other images, so it will probably break in more tricky cases!