Don't edge align product in crop if product not at border of image
In deciding whether or not to align a edge of the crop to the edge of
a product when the crop overflows the image:
Before only stopped this if the given dimension did not overflow. Now
add a check that the product is actually located at the edge in
question in the original image.
This prevents aligning the product strangely in the cropped photo.